+++ src/yb/yql/cql/ql/audit/audit_logger.h
@@ -54,7 +54,7 @@ class AuditLogger {
   // because in that case separate commands might arrive to different tservers.
   //
   // If this returns non-OK status, batch mode isn't activated.
-  CHECKED_STATUS StartBatchRequest(int statements_count,
+  CHECKED_STATUS StartBatchRequest(size_t statements_count,
                                    IsRescheduled is_rescheduled);
 
   // Exits the batch request mode. Does nothing outside of a batch request.
